# openapi.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AlexMubarakshin/ApiSchema/1.0.1
info:
  description: Api schema
  version: "1.0.1"
  title: API schema
security:
  - bearerAuth: []  
...
paths:
...
  /api/v1/users/me:
    get:
      tags:
        - profile
      description: "Получить текущего пользователя"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Успешное получение текущего пользователя"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: "#/components/schemas/Person"
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - profile
      description: "Редактирование текущего пользователя"
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonEditBody"
      responses:
        '200':
          description: "Успешное получение текущего пользователя"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: "#/components/schemas/Person"
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - profile
      description: "Удаление текущего пользователя"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Успешное удаление текущего пользователя"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      properties:
                        message:
                          type: string
                          description: "Сообщение о успехе"
                          example: "ok"
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/users/{id}/wall:
    get:
      tags:
        - profile
      description: "Получение записей на стене пользователя"
      parameters:
      - name: id
        in: path
        description: "ID пользователя"
        required: true
        schema:
          type: number
          example: 1
      - $ref: '#/components/parameters/offsetParam'
      - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: "Успешное получение публикаций"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/PersonsWallPost'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - profile
      description: "Добавление публикации на стену пользователя"
      security:
        - bearerAuth: []
      parameters:
      - name: id
        in: path
        description: "ID пользователя"
        required: true
        schema:
          type: number
          example: 1
      - name: publish_date
        in: query
        description: "Отложить до даты определенной даты"
        required: false
        schema:
          type: number
          example: 1559751301818
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequestBody'
      responses:
        '200':
          description: "Успешное добавление публикации"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Post'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/users/search/:
    get:
      tags:
        - profile
      description: "Поиск пользователя"
      summary: "Поиск пользователя"
      parameters:
      - name: first_name
        in: query
        description: "Имя пользователя"
        required: false
        schema:
          type: string
          example: "Григорий"
      - name: last_name
        in: query
        description: "Фамилия пользователя"
        required: false
        schema:
          type: string
          example: "Пепс"
      - name: age_from
        in: query
        description: "Кол-во лет ОТ"
        required: false
        schema:
          type: number
          example: 18
      - name: age_to
        in: query
        description: "Кол-во лет ДО"
        required: false
        schema:
          type: number
          example: 50
      - name: country_id
        in: query
        description: "ID страны"
        required: false
        schema:
          type: number
          example: 1
      - name: city_id
        in: query
        description: "ID города"
        required: false
        schema:
          type: number
          example: 3
      - $ref: '#/components/parameters/offsetParam'
      - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: "Успешное получение списка пользователей"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/Person'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/post/:
    get:
      tags:
        - posts
      description: "Поиск публикации"
      summary: "Поиск публикации"
      security:
        - bearerAuth: []
      parameters:
      - name: text
        in: query
        description: "Текст публикации"
        required: true
        schema:
          type: string
          example: "Бла бла"
      - name: date_from
        in: query
        description: "Дата публикации ОТ"
        required: false
        schema:
          type: number
          example: 1559751301818
      - name: date_to
        in: query
        description: "Дата публикации ДО"
        required: false
        schema:
          type: number
          example: 1559751301818
      - $ref: '#/components/parameters/offsetParam'
      - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: "Успешное получение публикации"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/Post'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/post/{id}:
    get:
      tags:
        - posts
      description: "Получение публикации по ID"
      summary: "Получение публикации по ID"
      security:
        - bearerAuth: []
      parameters:
      - name: id
        in: path
        description: "ID публикации"
        required: true
        schema:
          type: number
          example: 1
      responses:
        '200':
          description: "Успешное получение публикации"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Post'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - posts
      description: "Редактирование публикации"
      summary: "Редактирование публикации по ID"
      security:
        - bearerAuth: []
      parameters:
      - name: id
        in: path
        description: "ID публикации"
        required: true
        schema:
          type: number
          example: 1
      - name: publish_date
        in: query
        description: "Отложить до даты определенной даты"
        required: false
        schema:
          type: number
          example: 1559751301818
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequestBody'
      responses:
        '200':
          description: "Успешное редактирование публикации"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/Post'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - posts
      description: "Удаление публикации"
      summary: "Удаление публикации по ID"
      security:
        - bearerAuth: []
      parameters:
      - name: id
        in: path
        description: "ID публикации"
        required: true
        schema:
          type: number
          example: 1
      responses:
        '200':
          description: "Успешное удаление публикации"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/Response"
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        id:
                          type: number
                          description: "ID удаленной публикации"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/notifications:
    get:
      tags:
        - notifications
      description: "Получить список уведомлений для текущего пользователя"
      summary: "Получить список уведомлений"
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: "Успешное получение списка уведомлений"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/NotificationBase'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:  
      tags:
        - notifications
      description: "Отметить уведомление как \"прочитанное\""
      summary: "Прочитать уведомление"
      security:
        - bearerAuth: []
      parameters:
      - name: id
        in: query
        description: "ID уведомления"
        required: false
        schema:
          type: number
          example: 1
      - name: all
        in: query
        description: "Пометка что прочесть все уведомления"
        required: false
        schema:
          type: boolean
          example: true
        
      responses:
        "200":
          description: "Успешное получение списка уведомлений"
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/ListResponse"
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/NotificationBase'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: "unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Тип ошибки"
          example: "invalid_request"
          enum:
            - invalid_request
            - unauthorized
        error_description:
          type: string
          items:
            $ref: '#/components/schemas/ErrorDescription'
    ErrorDescription:
      type: object
      properties:
        error_description:
          type: string
          description: "Error description"
          enum:
            - "Unauthorized"
            - "An authorization code must be supplied"
            - "Redirect URI mismatch"
            - "Invalid authorization code: CODE"
            - "Bad credentials"
    Response:
      type: object
      properties:
        error: 
          type: string
          description: "Сообщение, в случае ошибки здесь будет причина"
        timestamp:
          type: number
          example: 1559751301818
    ListResponse:
      allOf:
      - $ref: "#/components/schemas/Response"
      - type: object
        properties:
          total:
            type: integer
            description: "Количество элементов в списке"
            format: int64
          offset:
            type: integer
            description: "Отступ от начала списка"
            format: int64
          perPage:
            type: integer
            description: "Количество элементов на страницу"
            format: int64
            default: 20
    PersonEditBody:
      type: object
      description: "Пользователь соц. сети"
      properties:
        first_name:
          type: string
          example: "Петр"
        last_name:
          type: string
          example: "Петрович"
        birth_date:
          type: number
          example: 1559751301818
        phone:
          type: string
          example: '89100000000'
        photo_id:
          type: number
          description: "ID на фото в хранилище"
          example: "o1doj1d91j1d01d-1d1f"
        about:
          type: string
          example: "Родился в небольшой, но честной семье"
        town_id:
          type: number
          example: 123
        country_id:
          type: number
          example: 33
        messages_permission:
          type: string
          example: "ALL"
          description: "Разрешение на получение сообщений: ALL - от всех пользователей (кроме заблокированных), FRIENDS - только от друзей"
    Person:
      type: object
      description: "Пользователь соц. сети"
      properties:
        id:
          type: number
          example: 1
        first_name:
          type: string
          example: "Петр"
        last_name:
          type: string
          example: "Петрович"
        reg_date:
          type: number
          example: 1559751301818
        birth_date:
          type: number
          example: 1559751301818
        email:
          type: string
          example: "petr@mail.ru"
        phone:
          type: string
          example: '89100000000'
        photo:
          type: string
          example: "https://...../photos/image123.jpg"
        about:
          type: string
          example: "Родился в небольшой, но честной семье"
        city:
          $ref: "#/components/schemas/City"
        country:
          $ref: "#/components/schemas/Country"
        messages_permission:
          type: string
          example: "ALL"
          description: "Разрешение на получение сообщений: ALL - от всех пользователей (кроме заблокированных), FRIENDS - только от друзей"
        last_online_time:
          type: number
          example: 1559751301818
        is_blocked:
          type: boolean
          example: false
...
    PostRequestBody:
      type: object
      properties:
        title:
          type: string
          description: "Заголовок"
        post_text:
          type: string
          description: "HTML-текст поста"
    Post:
      type: object
      properties:
        id:
          type: number
          example: 1
        time:
          type: number
          example: 1559751301818
          description: "дата и время публикации"
        author:
          $ref: "#/components/schemas/Person"
            
        title:
          type: string
          description: "Заголовок"
        post_text:
          type: string
          description: "HTML-текст поста"
        is_blocked:
          type: boolean
          description: "Отметка о том, что пост заблокирован"
          example: false
        likes:
          type: number
          description: "Количество лайков"
          example: 23
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    PersonsWallPost:
      description: "Пост на стене пользователя"
      allOf:
        - $ref: "#/components/schemas/Post"
        - type: object
          properties:
            type:
              type: string
              enum:
                - POSTED
                - QUEUED
              example: "POSTED"
...
    CommentModel:
      type: object
      description: "Комментарий к посту"
      properties:
        parent_id:
          type: number
          example: 1
          description: "Родительский комментарий (если ответ на комментарий к посту)"
        comment_text:
          type: string
    Comment:
      type: object
      description: "Комментарий к посту"
      allOf:
        - $ref: "#/components/schemas/CommentModel"
        - type: object
          properties:
            id:
              type: number
              description: "Id комментария"
              example: 111
            post_id:
              type: string
            time:
              type: number
              example: 1559751301818
            author_id:
              type: number
              example: 1
            is_blocked:
              type: boolean
              description: "Комментарий заблокирован"
    NotificationBase:
      type: object
      description: "Оповещения"
      properties:
        id:
          type: number
          example: 1
        type_id:
          type: number
          example: 1
          description: "ID типа оповещения"
        sent_time:
          type: number
          example: 1559751301818
          description: "Время отправки"
        entity_id:
          type: number
          example: 1
          description: "Идентификатор сущности, относительно которой отправлено оповещение (комментарий, друг, пост или сообщение)"
        info:
          type: string
          description: "Текст уведомления"
...
    Country:
      type: object
      properties:
        id:
          type: number
          description: "Id страны"
          example: 1
        title:
          type: string
          description: "Название страны"
          example: "Россия"
    City:
      type: object
      properties:
        id:
          type: number
          description: "Id города"
          example: 1
        title:
          type: string
          description: "Название города"
          example: "Москва"
  parameters:
    offsetParam:
      in: query
      description: "Отступ от начала списка"
      name: offset
      required: false
      schema:
        type: integer
        format: int64
    limitParam:
      in: query
      description: "Количество элементов на страницу"
      name: itemPerPage
      required: false
      schema:
        default: 20
        type: integer
        format: int64